+++ xivo_ctid_ng/config.py
@@ -107,6 +107,7 @@ _DEFAULT_CONFIG = {
         'calls': True,
         'chats': True,
         'conferences': True,
+        'faxes': True,
         'mongooseim': True,
         'presences': True,
         'relocates': True,